backend/apps/auth/utils.py
from datetime import datetime, timedelta
from jose import jwt, JWTError
from django.conf import settings
from django.contrib.auth import get_user_model

User = get_user_model()


def create_token(user_id: int) -> str:
    payload = {
        "user_id": user_id,
        "exp": datetime.utcnow() + timedelta(seconds=settings.JWT_EXP_DELTA_SECONDS),
    }
    return jwt.encode(payload, settings.JWT_SECRET, algorithm=settings.JWT_ALGORITHM)


def decode_token(token: str):
    try:
        payload = jwt.decode(
            token, settings.JWT_SECRET, algorithms=[settings.JWT_ALGORITHM]
        )
        return payload.get("user_id")
    except JWTError:
        return None


def get_user_from_token(token: str):
    user_id = decode_token(token)
    if user_id:
        try:
            return User.objects.get(id=user_id)
        except User.DoesNotExist:
            return None
    return None
