backend/requirements.txt
Django==4.2.11
django-ninja==1.1.0
psycopg2-binary==2.9.9
pydantic==2.6.3
pydantic[email]==2.6.3
python-jose[cryptography]==3.3.0
passlib==1.7.4
pytest==8.1.1
pytest-django==4.8.0
django-cors-headers==4.3.1